Fibonacci
The fibonacci sequence is defined here Links to an external site.. Write an implementation of the fibonacci sequence that computes the nth term of the sequence where n is a random number from 15 to 35 inclusive. Time the number of seconds your implementation takes with the Python time module Links to an external site.. Print out both the result of your script's fibonacci function and the time it took, as shown below.